% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{aic}
\alias{aic}
\title{AIC}
\usage{
aic(loglik, k)
}
\arguments{
\item{loglik}{A numeric, the log-likelihood.}

\item{k}{An integer >= 0, the number of selected variables.}
}
\value{
A number, a value of AIC.
}
\description{
Calculate AIC (Akaike Information Criterion).
}
\examples{
aic(10, 5)
}
