% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{maic}
\alias{maic}
\title{mAIC}
\usage{
maic(loglik, k, p, const = 0.5)
}
\arguments{
\item{loglik}{A numeric, the log-likelihood.}

\item{k}{An integer >= 0, the number of selected variables.}

\item{p}{An integer > 0, the number of all variables or a weight.}

\item{const}{A numeric > 0, the expected number of significant variables.}
}
\value{
A number, a value of mAIC.
}
\description{
Calculate mAIC (modified Akaike Information Criterion).
}
\examples{
maic(10, 5, 100, 50)
}
