% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{mbic2}
\alias{mbic2}
\title{mBIC2}
\usage{
mbic2(loglik, k, n, p, const = 4)
}
\arguments{
\item{loglik}{A numeric, the log-likelihood.}

\item{k}{An integer >= 0, the number of selected variables.}

\item{n}{An integer > 0, the number of observations.}

\item{p}{An integer > 0, the number of all variables or a weight.}

\item{const}{A numeric > 0, the expected number of significant variables.}
}
\value{
A number, a value of mBIC2.
}
\description{
Calculate mBIC2 (the second version of modified Bayesian Information
Criterion).
}
\examples{
mbic2(10, 5, 100, 50)
}
