% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{AIC}
\usage{
aic(rss, n, k)
}
\arguments{
\item{rss}{a numeric > 0, the residual sum of squares.}

\item{n}{an integer > 0, a number of observations.}

\item{k}{an integer >= 0, a number of selected variables.}
}
\value{
A number, a value of AIC.
}
\description{
Calculate AIC (Akaike Information Criterion).
}
\examples{
aic(rss=10, n=100, k=5)
}

