% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic}
\alias{bic}
\title{BIC}
\usage{
bic(loglik, n, k)
}
\arguments{
\item{loglik}{a numeric, the log-likelihood.}

\item{n}{an integer > 0, a number of observations.}

\item{k}{an integer >= 0, a number of selected variables.}
}
\value{
A number, a value of BIC.
}
\description{
Calculate BIC (Bayesian Information Criterion).
}
\examples{
bic(loglik=10, n=100, k=5)
}

