% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{maic}
\alias{maic}
\title{mAIC}
\usage{
maic(rss, n, k, p, const = 4)
}
\arguments{
\item{rss}{a numeric > 0, the residual sum of squares.}

\item{n}{an integer > 0, a number of observations.}

\item{k}{an integer >= 0, a number of selected variables.}

\item{p}{an integer > 0, a number of all variables or a weight.}

\item{const}{numeric > 0, the expected number of significant variables.}
}
\value{
A number, a value of mAIC.
}
\description{
Calculate mAIC (modified Akaike Information Criterion).
}
\examples{
maic(rss=10, n=100, k=5, p=50)
}

