% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic2.R
\name{maic2}
\alias{maic2}
\title{mAIC2}
\usage{
maic2(loglik, n, k, p, const = 4)
}
\arguments{
\item{loglik}{a numeric, the log-likelihood.}

\item{n}{an integer > 0, a number of observations.}

\item{k}{an integer >= 0, a number of selected variables.}

\item{p}{an integer > 0, a number of all variables or a weight.}

\item{const}{numeric > 0, the expected number of significant variables.}
}
\value{
A number, a value of mAIC2.
}
\description{
Calculate mAIC2 (the second version of modified Akaike Information
Criterion).
}
\examples{
mbic2(loglik=10, n=100, k=5, p=50)
}

