% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{maic2}
\alias{maic2}
\title{mAIC2}
\usage{
maic2(loglik, k, p, const = 4)
}
\arguments{
\item{loglik}{A numeric, the log-likelihood.}

\item{k}{An integer >= 0, the number of selected variables.}

\item{p}{An integer > 0, the number of all variables or a weight.}

\item{const}{A numeric > 0, the expected number of significant variables.}
}
\value{
A number, a value of mAIC2.
}
\description{
Calculate mAIC2 (the second version of modified Akaike Information
Criterion).
}
\examples{
maic2(10, 5, 100, 50)
}
