% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbic2.R
\name{mbic2}
\alias{mbic2}
\title{mBIC2}
\usage{
mbic2(loglik, n, k, p, const = 4)
}
\arguments{
\item{loglik}{a numeric, the log-likelihood.}

\item{n}{an integer > 0, a number of observations.}

\item{k}{an integer >= 0, a number of selected variables.}

\item{p}{an integer > 0, a number of all variables or a weight.}

\item{const}{numeric > 0, the expected number of significant variables.}
}
\value{
A number, a value of mBIC2.
}
\description{
Calculate mBIC2 (the second version of modified Bayesian Information
Criterion).
}
\examples{
mbic2(loglik=10, n=100, k=5, p=50)
}

