% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{read.big.matrix}
\alias{reexports}
\alias{attach.big.matrix}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bigmemory}{\code{\link[bigmemory]{read.big.matrix}}, \code{\link[bigmemory]{attach.big.matrix}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

